Added some more classes and A proto maven setup
The idea is to build a template rest environment in RESTfullApp to see how one works.  The idea of this move is to fix the frequent data cash errors generated when the user hits the back button at any point during the current app.  the gole is to create a data file and set all values to getters and setters that the code updates rather than the cash