Patch 6.0
Patch Notes 6.0
Did a major Overhaul of the Programs language fixing the print statement instructions and making sure everything in the search engine is spelled correctly.
Added slight cosmetic adjustments to the Display page
Added a Dynamic state to the Modal.jsp to change it's exit parameters depending on the path taken to open it. 
Added feature to search engine to disable select all button when there is only one entry. 
Finished the error handling for the search engine the code now looks out for the following missing fields
failure to select a table filed to search through
failure to type in a search
failure to spell your request correctly
failure to input the correct number of words
put in something other than a number value to search by
and expanded the excepted range of user input to include letters and symbols  and the dose not equal(!=) operations
Added the final page to the project and completed the data storage process
started the data repository for better data retention and manipulation.
made all of the "Byte streams" operations dynamic.
Finished the cosmetic changes to the report page.
Finished writing the dynamic bin creation method for the histogram as well as a sort array function.
Updated and expanded the Junit test cases