Patch Notes 7.0
Completed user Search engine Error handling.  Added a restart function to the Code to ease the need for constant resets.  Added a new page that had a collection of all the mode values and a "print to file" save feature.  Updated the server logs to include a copy of all printed texts. Added a Design mission statement to the front page. Fixed the Data cash problems, and added some REST API conversion pages.