Final build File
A complete website that combines all the major concepts explored by the other projects on this Git hub projects. 
Contains
Rest API, Crud, junit, spring, JDBC, Java, and Angular JS
A complete website build with roles data and Analytics
Missing a lot of the required CSS to make the sight look profeshinal.  Working on that soon to be update